single-source the DOM truth tables in src/dom-tables.js
The boolean/must-use-property attribute sets, attribute aliases, SVG-only
tag classification, unitless style props, and void elements were
hand-duplicated between compile.js, constants.ts, and the runtimes with
"keep in sync" comments (project-analysis-concerns.md §5) — drift would
surface as client/SSR/hydration disagreement. They now live only in
src/dom-tables.js, a plain-JS module the verbatim-shipped compiler can
import directly; constants.ts re-exports with type annotations so the
runtimes and the public octane/constants surface are unchanged. Same
Set/Map instances at every call site — no lookup-shape changes.

The member-level audit found one real drift, now fixed: the compiler's
private cssStyleValueStatic didn't trim string style values while the
runtimes' cssStyleValue does, so a statically-baked {color: ' red '}
diverged from the dynamic/SSR bytes. Static bakes share the runtime
function now.

Also fixes the publish build: css.ts, server/rpc.ts, and static/index.ts
were missing from the esbuild entryPoints, leaving dist/runtime.js,
octane/server, and octane/static with unresolvable imports.

tests/dom-tables.test.ts pins the wiring itself: re-export identity, a
mutate-the-shared-Set compiler probe, and per-category bake assertions.
