+++ packages/octane/src/compiler/compile.js
@@ -37,22 +37,20 @@ import {
 import { print as esrapPrint } from 'esrap';
 import esrapTsx from 'esrap/languages/tsx';
 
-const VOID_ELEMENTS = new Set([
-	'area',
-	'base',
-	'br',
-	'col',
-	'embed',
-	'hr',
-	'img',
-	'input',
-	'link',
-	'meta',
-	'param',
-	'source',
-	'track',
-	'wbr',
-]);
+// DOM truth tables shared with the client/server runtimes (via constants.ts) —
+// static bakes and dynamic writes MUST agree on which attributes render, under
+// what name, and in what form, or client/SSR/hydration drift apart. See the
+// dom-tables.js header for per-table semantics.
+import {
+	VOID_ELEMENTS,
+	BOOLEAN_ATTR_PROPS,
+	MUST_USE_PROPERTY_PROPS,
+	SVG_ONLY_TAGS,
+	ATTRIBUTE_ALIASES,
+	isEnumeratedBooleanAttr,
+	cssStyleValue,
+	hyphenateStyleName,
+} from '../dom-tables.js';
 
 // React parity: a void element must neither have children nor use
 // `dangerouslySetInnerHTML` — React throws (ReactDOMComponent-test.js:1794/:1807).
@@ -140,37 +138,6 @@ const CONTROLLED_KIND_HELPERS = {
 	autoFocus: 'setAutoFocus',
 };
 
-// React's boolean attribute props + the mustUseProperty set — DUPLICATED from
-// src/constants.ts (keep in sync): static literals bake the canonical
-// presence form (`disabled=""`) for truthy values and DROP falsy ones,
-// matching the runtimes' dynamic writes (coerceAttrValue / ssrAttr).
-const BOOLEAN_ATTR_PROPS = new Set([
-	'allowfullscreen',
-	'async',
-	'autoplay',
-	'controls',
-	'default',
-	'defer',
-	'disabled',
-	'disablepictureinpicture',
-	'disableremoteplayback',
-	'formnovalidate',
-	'hidden',
-	'inert',
-	'itemscope',
-	'loop',
-	'nomodule',
-	'novalidate',
-	'open',
-	'playsinline',
-	'readonly',
-	'required',
-	'reversed',
-	'scoped',
-	'seamless',
-]);
-const MUST_USE_PROPERTY_PROPS = new Set(['muted', 'multiple', 'selected']);
-
 // Serialize one STATIC literal attribute value into template/SSR HTML —
 // shared by emitElementHtml and ssrEmitElement so both bakes stay identical,
 // mirroring the runtimes' dynamic coercion: aria-*/enumerated/data-* booleans
@@ -187,8 +154,8 @@ function bakeStaticAttr(attrName, lv, tag) {
 			return ` ${attrName}="${lv}"`;
 		}
 		// Enumerated booleans stringify — "false" is a real state, absent means
-		// inherit (mirrors isEnumeratedBooleanAttr in the runtimes).
-		if (lower === 'spellcheck' || lower === 'draggable' || lower === 'contenteditable') {
+		// inherit (the same gate the runtimes apply to dynamic values).
+		if (isEnumeratedBooleanAttr(lower)) {
 			return ` ${attrName}="${lv}"`;
 		}
 		if (isCustom) return lv ? ` ${attrName}` : '';
@@ -380,84 +347,13 @@ export const HOOK_NAMES = new Set([
 // Namespace inheritance — mirrors HTML5 foreign-content rules. The element
 // itself and its children may have *different* namespaces: <foreignObject>
 // inside SVG is still an SVG element, but its children switch back to HTML.
-// Tags that exist ONLY in the SVG namespace, so their appearance in a
-// namespace-ambiguous position — a component's ROOT template, a fragment root
-// — implies the SVG namespace without a lexical `<svg>` ancestor. A component
-// whose root is `<g>`/`<path>` must not compile to an HTML template: the
-// HTMLUnknownElement it would produce inside an `<svg>` paints nothing.
-// Ambiguous names (`a`, `title`, `script`, `style`, `font`) are deliberately
-// ABSENT — they keep the inherited namespace. Keep in sync with SVG_ONLY_TAGS
-// in ../constants.ts (the runtime's de-opt reconciler uses the same table).
-const SVG_ONLY_TAGS = new Set([
-	'altGlyph',
-	'altGlyphDef',
-	'altGlyphItem',
-	'animate',
-	'animateColor',
-	'animateMotion',
-	'animateTransform',
-	'circle',
-	'clipPath',
-	'defs',
-	'desc',
-	'ellipse',
-	'feBlend',
-	'feColorMatrix',
-	'feComponentTransfer',
-	'feComposite',
-	'feConvolveMatrix',
-	'feDiffuseLighting',
-	'feDisplacementMap',
-	'feDistantLight',
-	'feDropShadow',
-	'feFlood',
-	'feFuncA',
-	'feFuncB',
-	'feFuncG',
-	'feFuncR',
-	'feGaussianBlur',
-	'feImage',
-	'feMerge',
-	'feMergeNode',
-	'feMorphology',
-	'feOffset',
-	'fePointLight',
-	'feSpecularLighting',
-	'feSpotLight',
-	'feTile',
-	'feTurbulence',
-	'filter',
-	'foreignObject',
-	'g',
-	'glyph',
-	'glyphRef',
-	'hkern',
-	'image',
-	'line',
-	'linearGradient',
-	'marker',
-	'mask',
-	'metadata',
-	'missing-glyph',
-	'mpath',
-	'path',
-	'pattern',
-	'polygon',
-	'polyline',
-	'radialGradient',
-	'rect',
-	'set',
-	'stop',
-	'switch',
-	'symbol',
-	'text',
-	'textPath',
-	'tref',
-	'tspan',
-	'use',
-	'view',
-	'vkern',
-]);
+// SVG_ONLY_TAGS (imported from ../dom-tables.js — the runtime's de-opt
+// reconciler uses the same table) drives the inference: a tag that exists ONLY
+// in the SVG namespace implies SVG in a namespace-ambiguous position — a
+// component's ROOT template, a fragment root — without a lexical `<svg>`
+// ancestor. A component whose root is `<g>`/`<path>` must not compile to an
+// HTML template: the HTMLUnknownElement it would produce inside an `<svg>`
+// paints nothing.
 
 function nsForSelf(tag, parentNs) {
 	if (tag === 'svg') return 'svg';
@@ -514,106 +410,11 @@ function jsxAttrRawName(attr) {
 	return n.name || n;
 }
 
-// React 19's attribute-alias table — camelCase JSX prop → the attribute the
-// browser actually parses (`strokeWidth` → `stroke-width`, `htmlFor` → `for`,
-// `xlinkHref` → `xlink:href`). An ALLOWLIST, not mechanical hyphenation
-// (`viewBox` stays camelCase). Keep in sync with ATTRIBUTE_ALIASES in
-// ../constants.ts — the compiler bakes these into static template/SSR markup,
-// the client/server runtimes apply the same table to dynamic bindings,
-// spreads, and de-opt props.
-const ATTRIBUTE_ALIASES = new Map([
-	['acceptCharset', 'accept-charset'],
-	['htmlFor', 'for'],
-	['httpEquiv', 'http-equiv'],
-	['crossOrigin', 'crossorigin'],
-	['accentHeight', 'accent-height'],
-	['alignmentBaseline', 'alignment-baseline'],
-	['arabicForm', 'arabic-form'],
-	['baselineShift', 'baseline-shift'],
-	['capHeight', 'cap-height'],
-	['clipPath', 'clip-path'],
-	['clipRule', 'clip-rule'],
-	['colorInterpolation', 'color-interpolation'],
-	['colorInterpolationFilters', 'color-interpolation-filters'],
-	['colorProfile', 'color-profile'],
-	['colorRendering', 'color-rendering'],
-	['dominantBaseline', 'dominant-baseline'],
-	['enableBackground', 'enable-background'],
-	['fillOpacity', 'fill-opacity'],
-	['fillRule', 'fill-rule'],
-	['floodColor', 'flood-color'],
-	['floodOpacity', 'flood-opacity'],
-	['fontFamily', 'font-family'],
-	['fontSize', 'font-size'],
-	['fontSizeAdjust', 'font-size-adjust'],
-	['fontStretch', 'font-stretch'],
-	['fontStyle', 'font-style'],
-	['fontVariant', 'font-variant'],
-	['fontWeight', 'font-weight'],
-	['glyphName', 'glyph-name'],
-	['glyphOrientationHorizontal', 'glyph-orientation-horizontal'],
-	['glyphOrientationVertical', 'glyph-orientation-vertical'],
-	['horizAdvX', 'horiz-adv-x'],
-	['horizOriginX', 'horiz-origin-x'],
-	['imageRendering', 'image-rendering'],
-	['letterSpacing', 'letter-spacing'],
-	['lightingColor', 'lighting-color'],
-	['markerEnd', 'marker-end'],
-	['markerMid', 'marker-mid'],
-	['markerStart', 'marker-start'],
-	['overlinePosition', 'overline-position'],
-	['overlineThickness', 'overline-thickness'],
-	['paintOrder', 'paint-order'],
-	['panose-1', 'panose-1'],
-	['pointerEvents', 'pointer-events'],
-	['renderingIntent', 'rendering-intent'],
-	['shapeRendering', 'shape-rendering'],
-	['stopColor', 'stop-color'],
-	['stopOpacity', 'stop-opacity'],
-	['strikethroughPosition', 'strikethrough-position'],
-	['strikethroughThickness', 'strikethrough-thickness'],
-	['strokeDasharray', 'stroke-dasharray'],
-	['strokeDashoffset', 'stroke-dashoffset'],
-	['strokeLinecap', 'stroke-linecap'],
-	['strokeLinejoin', 'stroke-linejoin'],
-	['strokeMiterlimit', 'stroke-miterlimit'],
-	['strokeOpacity', 'stroke-opacity'],
-	['strokeWidth', 'stroke-width'],
-	['textAnchor', 'text-anchor'],
-	['textDecoration', 'text-decoration'],
-	['textRendering', 'text-rendering'],
-	['transformOrigin', 'transform-origin'],
-	['underlinePosition', 'underline-position'],
-	['underlineThickness', 'underline-thickness'],
-	['unicodeBidi', 'unicode-bidi'],
-	['unicodeRange', 'unicode-range'],
-	['unitsPerEm', 'units-per-em'],
-	['vAlphabetic', 'v-alphabetic'],
-	['vHanging', 'v-hanging'],
-	['vIdeographic', 'v-ideographic'],
-	['vMathematical', 'v-mathematical'],
-	['vectorEffect', 'vector-effect'],
-	['vertAdvY', 'vert-adv-y'],
-	['vertOriginX', 'vert-origin-x'],
-	['vertOriginY', 'vert-origin-y'],
-	['wordSpacing', 'word-spacing'],
-	['writingMode', 'writing-mode'],
-	['xmlnsXlink', 'xmlns:xlink'],
-	['xHeight', 'x-height'],
-	['xlinkActuate', 'xlink:actuate'],
-	['xlinkArcrole', 'xlink:arcrole'],
-	['xlinkHref', 'xlink:href'],
-	['xlinkRole', 'xlink:role'],
-	['xlinkShow', 'xlink:show'],
-	['xlinkTitle', 'xlink:title'],
-	['xlinkType', 'xlink:type'],
-	['xmlBase', 'xml:base'],
-	['xmlLang', 'xml:lang'],
-	['xmlSpace', 'xml:space'],
-	// React writes this via a setProp switch case rather than its aliases map;
-	// same observable output. Matters on SVG hosts (setAttribute preserves case).
-	['tabIndex', 'tabindex'],
-]);
+// ATTRIBUTE_ALIASES (imported from ../dom-tables.js): React 19's alias table,
+// camelCase JSX prop → the attribute the browser actually parses
+// (`strokeWidth` → `stroke-width`, `htmlFor` → `for`). The compiler bakes
+// these into static template/SSR markup; the client/server runtimes apply the
+// same table to dynamic bindings, spreads, and de-opt props.
 
 // `className` plus the ATTRIBUTE_ALIASES table above: emit the native
 // attribute names so the browser actually applies them (and dynamic bindings
@@ -660,102 +461,18 @@ function objectExprIsStaticLiteral(obj) {
 	return true;
 }
 
-// camelCase / vendor-prefixed style key → kebab-case. Mirrors the runtime
-// `styleName` (runtime.ts) / `hyphenate` (runtime.server.ts) so a STATIC baked
-// object style produces the same CSS a dynamic one would.
-function hyphenateStyleName(name) {
-	if (name.charCodeAt(0) === 45 /* - */) return name; // --custom / -webkit-…
-	let out = '';
-	let changed = false;
-	for (let i = 0; i < name.length; i++) {
-		const c = name.charCodeAt(i);
-		if (c >= 65 && c <= 90) {
-			out += '-' + String.fromCharCode(c + 32);
-			changed = true;
-		} else out += name[i];
-	}
-	if (!changed) return name;
-	if (out.charCodeAt(0) === 109 && out.charCodeAt(1) === 115 && out.charCodeAt(2) === 45)
-		out = '-' + out;
-	return out;
-}
-
-// React's unitless-property set — keep in sync with `constants.ts`
-// (UNITLESS_STYLE_PROPS). Canonical form: lowercased, dashes stripped.
-const UNITLESS_STYLE_PROPS = new Set();
-for (const base of [
-	'animationIterationCount',
-	'aspectRatio',
-	'borderImageOutset',
-	'borderImageSlice',
-	'borderImageWidth',
-	'boxFlex',
-	'boxFlexGroup',
-	'boxOrdinalGroup',
-	'columnCount',
-	'columns',
-	'flex',
-	'flexGrow',
-	'flexPositive',
-	'flexShrink',
-	'flexNegative',
-	'flexOrder',
-	'gridArea',
-	'gridRow',
-	'gridRowEnd',
-	'gridRowSpan',
-	'gridRowStart',
-	'gridColumn',
-	'gridColumnEnd',
-	'gridColumnSpan',
-	'gridColumnStart',
-	'fontWeight',
-	'lineClamp',
-	'lineHeight',
-	'opacity',
-	'order',
-	'orphans',
-	'tabSize',
-	'widows',
-	'zIndex',
-	'zoom',
-	'fillOpacity',
-	'floodOpacity',
-	'stopOpacity',
-	'strokeDasharray',
-	'strokeDashoffset',
-	'strokeMiterlimit',
-	'strokeOpacity',
-	'strokeWidth',
-]) {
-	const c = base.toLowerCase();
-	UNITLESS_STYLE_PROPS.add(c);
-	UNITLESS_STYLE_PROPS.add('webkit' + c);
-	UNITLESS_STYLE_PROPS.add('ms' + c);
-	UNITLESS_STYLE_PROPS.add('moz' + c);
-	UNITLESS_STYLE_PROPS.add('o' + c);
-}
-
-// React parity: a bare number gets `px` unless it's 0, a custom prop, or unitless.
-function cssStyleValueStatic(name, value) {
-	if (
-		typeof value === 'number' &&
-		value !== 0 &&
-		name.charCodeAt(0) !== 45 &&
-		!UNITLESS_STYLE_PROPS.has(name.replaceAll('-', '').toLowerCase())
-	) {
-		return value + 'px';
-	}
-	return '' + value;
-}
-
+// Serialize a fully-literal style object into a `style="…"` string at compile
+// time. `hyphenateStyleName` + `cssStyleValue` come from ../dom-tables.js —
+// the SAME key normalization and px/unitless/trim coercion the runtimes apply
+// to dynamic style objects — so a STATIC baked object style produces the same
+// CSS a dynamic one would.
 function staticObjectToCssString(obj) {
 	const parts = [];
 	for (const p of obj.properties || []) {
 		const name = p.key.type === 'Identifier' ? p.key.name : p.key.value;
 		const value = p.value.value;
 		if (value == null || value === false || value === '') continue;
-		const cssValue = value === true ? '' : cssStyleValueStatic(name, value);
+		const cssValue = value === true ? '' : cssStyleValue(name, value);
 		parts.push(`${hyphenateStyleName(name)}: ${cssValue}`);
 	}
 	return parts.join('; ');
